Generate Join Expression with as available in query
The purpose of CXformExpandNaryJoin is to create a join tree with the
same order of joins as in the query.  The current logic however does not
respect the join order. It takes the scalar conditions existing in the
query and starts building the join.  For example: select * from t1, t2,
t3 where t1.a = t3.a; For the above query, the join produced will be (t1
join t3) join t2, however this is not the join order in the query.
Originally, before v2.50.2 the CXformExpandNaryJoin generated both the
alternatives, the one shown above and the alternative with the join
order in query (t1 join t2) join t3.

I don't see any reason to keep the logic to generate the join based on
the order of scalar conditions, because it's not the job of the
CXformExpandNaryJoin to explore that alternative. DP Xform considers all
the alternatives and retains the top 10 based on the cost. So, if its
cheaper, it will be considered by DP.

Signed-off-by: Melanie Plageman<m.plageman@pivotal.io>