Avoid starting timer unnecessarily
CTimerUser default constructor calls Restart() method which calls getrusage
which is a costly system call. We use the timer to print optimization
statistics if the trace flag EopttracePrintOptimizationStatistics is
set. Since the default constructor calls Restart(), even though
optimization statistics trace is off, we still pay cost to get the
default container. This patch ensures that default constructor of
CTimerUser does not call restart, if required Restart should be called
on CTimerUser object.

In ORCA, there is currently one search stage and the time out threshold
is set to ULONG_MAX which essentially means a timeout after more than a
month, so essentially no timeout is forced. So, there is no need to set
the timer in CSearchStage::RestartTimer which calls Restart() when the
time threshold is set to ULONG_MAX and similarly for
CSearchStage::FTimedOut which checks if the search stage has timed out.