Add cardinality estimation based on NDV for preds of form f(a) = b
Previously, predicates of the form f(a) = b were considered unsupported
and the cardinality for a join having such a predicate was
estimated using a constant scale factor. This resulted in gross
overestimation, and, in some cases, choosing an inferior plan.

With this commit, predicates of this form on an equijoin with exactly
one scalar ident on either side will be considered supported. This
commit introduces a new comparison type which will be used for
predicates like this and for which we estimate cardinality according to
the following formula: min(NDV(a), NDV(b)).