Fix logic for calculating number of times the Xform is applied
In the existing logic, for a Group Expression, the number of times the
Xform is applied is marked as 1, however, this is incorrect.
For a Group Expression, there may be several expression which match the
pattern defined by Xform, and the Xform will be applied on all the
patterns. This commit fixes the calculation logic.