Exclude BCC during Join Cardinality Estimation
While calculating the cardinality of a node, the implied predicates should
be removed as it will result in underestimating the cardinality. If the
cardinality estimates are row, optimizer may decide to place a broadcast
motion which can significantly reduce the performance of the query. This
commit fixes the cardinality under-estimation by not considering the
implied predictates during estimation.