Introduce a new scale factor algorithm for joins
This commit introduces a new algorithm for calculating the cumulative
scale factor for joins when estimating cardinality. The new algorithm is
partially based off of SQLServer2014 which uses moderately decreases the
impact of subsequent predicates by a sqrt factor.

The new algorithm only damps the predicate if there are multiple
predicates on the same two tables. Otherwise, we assume that there is no
correlation between predicates and thus don't need to be damped.

For example, if given the following join predicates:
    t1.a = t2.a AND t1.b = t2.b AND t2.b = t3.a

The algorithm would do the following:
    cumulative scale factor = (SF of t1.a = t2.a) * sqrt(SF of t1.b =
t2.b) * (SF of t2.b = t3.a)

Co-authored-by: Ashuka Xue <axue@pivotal.io>
Co-authored-by: Chris Hajas <chajas@pivotal.io>